refactor: moved artifact upload quick pick form to sep. file
