+++ src/commands/flinkArtifacts.test.ts
@@ -21,6 +21,7 @@ import {
   uploadArtifactCommand,
 } from "./flinkArtifacts";
 import * as commands from "./index";
+import * as artifactUploadForm from "./utils/artifactUploadForm";
 import * as uploadArtifact from "./utils/uploadArtifactOrUDF";
 
 describe("flinkArtifacts", () => {
@@ -136,7 +137,7 @@ describe("flinkArtifacts", () => {
   });
 
   it("should fail if there is no params", async () => {
-    sandbox.stub(uploadArtifact, "promptForArtifactUploadParams").resolves(undefined);
+    sandbox.stub(artifactUploadForm, "artifactUploadQuickPickForm").resolves(undefined);
     const result = await uploadArtifactCommand();
 
     assert.strictEqual(result, undefined);
@@ -150,7 +151,7 @@ describe("flinkArtifacts", () => {
       environment: " env-123456",
     };
 
-    sandbox.stub(uploadArtifact, "promptForArtifactUploadParams").resolves(mockParams);
+    sandbox.stub(artifactUploadForm, "artifactUploadQuickPickForm").resolves(mockParams);
     sandbox.stub(uploadArtifact, "getPresignedUploadUrl").resolves(mockPresignedUrlResponse);
     sandbox.stub(uploadArtifact, "handleUploadToCloudProvider").resolves();
     sandbox.stub(uploadArtifact, "uploadArtifactToCCloud").resolves(mockCreateResponse);
@@ -167,7 +168,7 @@ describe("flinkArtifacts", () => {
     const params = { ...mockParams };
     const uploadUrl = { ...mockPresignedUrlResponse };
 
-    sandbox.stub(uploadArtifact, "promptForArtifactUploadParams").resolves(params);
+    sandbox.stub(artifactUploadForm, "artifactUploadQuickPickForm").resolves(params);
     sandbox.stub(uploadArtifact, "getPresignedUploadUrl").resolves(uploadUrl);
     sandbox.stub(uploadArtifact, "handleUploadToCloudProvider").resolves();
 
@@ -194,7 +195,7 @@ describe("flinkArtifacts", () => {
       cloud: "Azure",
     };
 
-    sandbox.stub(uploadArtifact, "promptForArtifactUploadParams").resolves(mockParams);
+    sandbox.stub(artifactUploadForm, "artifactUploadQuickPickForm").resolves(mockParams);
     sandbox.stub(uploadArtifact, "getPresignedUploadUrl").resolves(mockPresignedUrlResponse);
     const handleUploadStub = sandbox.stub(uploadArtifact, "handleUploadToCloudProvider").resolves();
     const createArtifactStub = sandbox
